fix(swingset): addEgress should cause an import/reachable refcount
`addEgress` was calling `setReachable` with the wrong value for
`isImportFromComms`. When we add an egress, the kernel is exporting an object
into comms, and comms is exporting that to the downstream machine. So the
downstream machine is *importing* it from comms (which means the
`isReachable` flag causes the `reachable` refcount to be increased, something
that only happens on imports, not on the export).

This caused the object being exported through `addEgress` to have a zero
refcount. If/when the downstream machine ever dropped it, the refcount would
go negative, causing an error.

fixes #3483