fix(swingset): processRefcounts() even if crank was aborted
The kernelKeeper uses a non-durable `Set` named `maybeFreeKrefs` to track
krefs whose refcounts have touched zero. `kernelKeeper.processRefcounts()` is
then used to scan these and build any necessary GC Actions.

The old kernel code avoided performing this step if the crank was aborted,
e.g. if it provoked a vat-fatal syscall or metering fault. In that case, we
unwind the execution consequences (queued outbound messages, c-list changes,
etc). However we *do* commit the consequences of vat termination, like the
deletion of the c-lists.

The problem is that c-list deletion causes its own set of GC changes, since
the now-dead vat is no longer importing anything. If we don't do
`processRefcounts()` now, we're leaving the `maybeFreeKrefs` for the next
crank to handle. And since those are not durable, if the kernel process is
restarted, that work will be lost, and the dead vat's imports will be left
pinned unless/until something else touches their refcounts. Worse, the
resulting kernel state will not be the same as a kernel that did not
experience a restart, which would cause a consensus break.

This moves the `processRefcounts()` outside of the `if (!didAbort)` clause,
so it will be called unconditionally.

It also calls `saveStats()` unconditionally. I'm less confident about this
change. kernelKeeper holds some counters in RAM, ones which are changed at a
relatively high rate, for which it felt inefficient to touch the DB each
time. `saveStats()` flushes this write-back cache to the DB. If the delivery
is unwound, these stats may be inaccurate. But if we don't flush them now,
they'll be flushed at the end of the next crank, and the same "but what about
restart" / consensus-break argument applies. My current belief is that this
change might incorrectly count the stats, but at least all validators will
miscount them the same way.