fix(swingset-runner): remove --meter controls
Now that dynamic vats require a Meter object in the `meter:` option, rather
than just a `metered:` boolean flag, the swingset-runniner `--meter` option
doesn't work anymore. I'm removing it to allow the test to keep passing.
Eventually it might be interesting to improve the CLI arguments to create and
provide a Meter, but the need for user-level code to supervise and refill it
makes that non-trivial.