fix(swingset): use better async style, improve comment
The use of `await` inside a try/finally block would expose a subtle timing
issue (if `thunk()` throws synchronously, the finalizer would run in the same
turn as the thunk, rather than in a subsequent turn). Our recommended style
avoids conditional `await` for just that reason, so this updates
`runWithoutMeteringAsync` to use a better approach.