fix(swingset): comms initialization check must be deterministic
The comms vat has a bunch of counters that need to be initialized exactly
once in the lifetime of the vat's durable state. There is a DB flag named
`initialize` to track whether this has happened already or not. Since vats
can only do syscalls during deliveries, not startup (see #2910), the comms
vat must read this flag on every delivery, just in case this is the very
first one, and the DB needs initialization.

The comms vat was using an in-RAM flag (`needToInitializeState`) to cache the
result, to avoid a DB read for every single delivery. However this caused the
syscall behavior to change for the first delivery after a kernel restart.
There were two extra syscalls taking place: a `vatstoreGet('initialized')`,
and a `vatstoreSet('meta.o+0', 'true')`. The `vatstoreSet` was causing a DB
change, which was picked up by the new kernel activityhash, and caused a
consensus failure on the first post-restart comms message.

This changes the comms vat to always read the DB flag, on every delivery,
making its behavior consistent and independent of kernel restarts. It also
moves the `vatstoreSet` to be guarded by the results of the `initialized`
check, so it only happens once in the lifetime of the vat.

To remove the need for the `vatstoreGet` on each delivery, we'll need to
enhance the way `setup()` is called, to provide an external flag that tells
the vat whether this is the very first time ever, or if it's merely a kernel
restart that required the vat's `dispatch` object to be reconstructed.

fixes #3726