fix(xsnap)!: upgrade XS to fix memory leak
We upgrade the XS submodule to the latest version:
https://github.com/Moddable-OpenSource/moddable/commit/10cc52e130a62d59f9c13f5037c07bb32aeb64f2

This fixes a major memory leak: 64 bytes per Map `delete()`, 32 per Set
`delete()`. We believe this should: closes #3839

Unfortunately Map/Set deletion is now O(N) not O(1).

This version of XS also fixes a bug that might be the cause of #3877 "cannot
read (corrupted?) snapshot", but we're not sure.

Note that this breaks snapshot compatibility (snapshots created before this
version cannot be loaded by code after this version). It might also
break metering compatibility, but the chances seem low enough that we decided
to leave the metering version alone.

closes #3889