chore(xsnap): improve currentHeapCount in metering results
We landed the xsnap-native changes to add this field three years ago,
but while cleaning up old branches, I found a test that should go
along with it. Also an update to the types, and code to supply a null
value if the delivery did not include any metering in the response.

refs #3910

and
https://github.com/Agoric/agoric-sdk/issues/3910#issuecomment-938045207
in particular