drop internalCreatorFacet from contract governance (#4164)
* refactor: drop internalCreatorFacet from contract governance

I added the internalCreatorFacet as a result of my confusion.

The intent was that the limitedCreatorFacet would be the creatorFacet
provided by the contract with access to the contractGovernor added. I
forgot to interpolate the methods of the original creatorFacet into
the limitedCreatorFacet.

When I needed access to those methods later, I added the
internalCreatorFacet.

I've updated the contractGovernance diagram to show the methods.

contractGovernor starts the governedContract, and gets the powerful
creatorFacet. It uses the powerful methods itself, and passes only the
limited facet on to its caller.

* docs: clarify some governance requirements

a comment in contractGovernor said governed contract in place of contractGovernor

in contractHelper.js add API doc saying that the governed contract
shouldn't use the powerful facets

in AMM: reduce scope of creatorFacet by creating it after publicFacet

also some other cleanups
 * rename make*Facet to wrap*Facet

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>