chore(swingset): update docs on liveslots' dispatch() function
Chip's work on #2910 discovered that the supervisor was not told about
failures during liveslot's dispatch(). This could conceal some bugs in
liveslots, as well as hiding userspace-caused failures during
`buildRootObject()`.

The contract between `dispatch()` and the calling supervisor code has changed
through various bouts of refactoring, and it was ambiguous as to whether
`dispatch()` was supposed to protect against userspace errors or not. This
commit clears up the documentation to make this more explicit.